Added gov.nci.WebAnalytics.Resize
