+++ src/main/java/gov/nci/WebAnalytics/Resize.java
@@ -3,16 +3,14 @@ package gov.nci.WebAnalytics;
 import java.net.MalformedURLException;
 import java.util.List;
 import com.relevantcodes.extentreports.LogStatus;
+import org.openqa.selenium.Dimension;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
 import com.nci.testcases.AnalyticsTest;
 
-public class MegaMenu extends AnalyticsTest {
-
-
+public class Resize extends AnalyticsTest {
 	
-	//region browseractions
 	/**
 	 * All the proxy browser 'actions' go in here. These are not tests, but things that we do 
 	 * to fire off analytics events. These actions will populate our list of har objects, which will
@@ -20,22 +18,22 @@ public class MegaMenu extends AnalyticsTest {
 	 * @throws RuntimeException
 	 */
 	private void doBrowserActions() throws RuntimeException {
-		navigateSite();
+		resizeBrowser();
 	}
 	
-	/// Click around pages
-	public void navigateSite() {
+	// Resize browser
+	public void resizeBrowser() {
+		Dimension small = new Dimension(300, 800);
+		Dimension med = new Dimension(700, 800);
+		Dimension large = new Dimension(1100, 800);
+		Dimension xlarge = new Dimension(1600, 800);
 				
-		// Click on a feature card
-		clickEvents.clickFeatureCard();
-		driver.navigate().back();
-		
-		// Click on the MegaMenu
-		clickEvents.clickMegaMenu();		
-		driver.navigate().back();
-		
+		driver.manage().window().setSize(xlarge);
+		driver.manage().window().setSize(large);		
+		driver.manage().window().setSize(med);
+		driver.manage().window().setSize(small);
+		driver.manage().window().maximize();
 	}
-
 	//endregion browseractions
 	
 	//region tests
@@ -45,11 +43,9 @@ public class MegaMenu extends AnalyticsTest {
 	public void verifyHar() {
 		doBrowserActions();
 		List<String> harList = AnalyticsBase.getHarUrlList(proxy);
-		List<AnalyticsLoad> loadBeacons = AnalyticsLoad.getLoadBeacons(harList);
 		List<AnalyticsClick> clickBeacons = AnalyticsClick.getClickBeacons(harList);		
 				
 		Assert.assertTrue(harList.size() > 0);
-		Assert.assertTrue(loadBeacons.size() > 0);
 		Assert.assertTrue(clickBeacons.size() > 0);
 		
 		System.out.println("=== Start debug testEvents() ===");
@@ -59,26 +55,23 @@ public class MegaMenu extends AnalyticsTest {
 		System.out.println("=== End debug testEvents() ===");				
 		
 		logger.log(LogStatus.PASS, "Load and click events have been captured.");				
-	}	
+	}		
 	
-	/// Click event numbers match with their descriptors
+	/// Resize events match with their descriptors
 	@Test(groups = { "Analytics" })
-	public void testClickEvents() {
-		navigateSite();
+	public void testResizeEvents() {
+		resizeBrowser();
 		List<String> harList = AnalyticsBase.getHarUrlList(proxy);
 		List<AnalyticsClick> clickBeacons = AnalyticsClick.getClickBeacons(harList);
 		
 		for(AnalyticsClick beacon : clickBeacons) {
-			if(beacon.linkName == "FeatureCardClick") {
-				Assert.assertTrue(beacon.events[0].contains("event27"));
-			}
-			if(beacon.linkName == "MegaMenuClick") {
-				Assert.assertTrue(beacon.events[0].contains("event27"));
+			if(beacon.linkName.toLowerCase().contains("resize")) {
+				Assert.assertTrue(beacon.events[0].contains("event7"));
 			}
 		}
-		
-		logger.log(LogStatus.PASS, "Click event values are correct.");		
-	}		
+		logger.log(LogStatus.PASS, "Resize values are correct.");
+	}
+	
 	
 	/// Temporary method to verify that my new changes are picked up
 	@Test(groups = { "Analytics" })
